+++ flatlaf-swingx/src/main/java/com/formdev/flatlaf/swingx/icons/FlatMonthUpIcon.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.formdev.flatlaf.swingx.ui;
+package com.formdev.flatlaf.swingx.icons;
 
 import javax.swing.SwingConstants;